requirements.txt
# Unified XAI Platform - Requirements
# Python 3.8+ recommended

# Core Framework
streamlit>=1.28.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# XAI Libraries
lime>=0.2.0.1
shap>=0.42.0

# Data Processing
numpy>=1.23.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
